Implement bytecode refactor spec in circuit (#1044)
Closes #1026
