+++ zkevm-circuits/src/bytecode_circuit.rs
@@ -2,6 +2,8 @@
 
 /// Bytecode unroller
 pub mod bytecode_unroller;
+/// Bytecode circuit
+pub mod circuit;
 /// Bytecode circuit tester
 #[cfg(any(feature = "test", test))]
 pub mod dev;
