+++ zkevm-circuits/src/table.rs
@@ -552,12 +552,10 @@ impl MptTable {
 /// Tag to identify the field in a Bytecode Table row
 #[derive(Clone, Copy, Debug)]
 pub enum BytecodeFieldTag {
-    /// Length field
-    Length,
+    /// Header field
+    Header,
     /// Byte field
     Byte,
-    /// Padding field
-    Padding,
 }
 impl_expr!(BytecodeFieldTag);
 
@@ -839,6 +837,21 @@ impl KeccakTable {
                 Ok(())
             },
         )
+    }
+
+    /// returns matchings between the circuit columns passed as parameters and
+    /// the table collumns
+    pub fn match_columns(
+        &self,
+        value_rlc: Column<Advice>,
+        length: Column<Advice>,
+        code_hash: Column<Advice>,
+    ) -> Vec<(Column<Advice>, Column<Advice>)> {
+        vec![
+            (value_rlc, self.input_rlc),
+            (length, self.input_len),
+            (code_hash, self.output_rlc),
+        ]
     }
 }
